[SELC-5918] Feat: Hide sidebar for add role page + refactor names of files
