+++ src/utils/test-utils.tsx
@@ -9,8 +9,8 @@ export const renderWithProviders = (
   injectedStore?: ReturnType<typeof createStore>,
   injectedHistory?: ReturnType<typeof createMemoryHistory>
 ) => {
-  const store = injectedStore ? injectedStore : createStore();
-  const history = injectedHistory ? injectedHistory : createMemoryHistory();
+  const store = injectedStore || createStore();
+  const history = injectedHistory || createMemoryHistory();
   render(
     <Router history={history}>
       <Provider store={store}>{component}</Provider>
